Weighing finally works
In the end, had to resort to how weigh does it: by running the
executable in a special way for each function to be weighed.

It seems that it isn't possible for GHC to forget that it's run this
at least once for testing purposes.

Compared to weigh, testbench tries to recover by just returning
"Nothing" if a test-case fails for some reason; we need to work out
some way of indicating the various types of failure (program failed vs
program returned Nothing vs "That particular function doesn't
exist").  We also don't require unique labels.

See https://github.com/fpco/weigh/issues/10